+++ src/loss.cc
@@ -6,23 +6,23 @@
 
 #include "ndarray.h"
 
-float SoftmaxLoss(const Ndarray& x, const std::vector<int64_t>& y,
+double SoftmaxLoss(const Ndarray& x, const std::vector<int64_t>& y,
                   Ndarray* dx) {
   assert(dx != nullptr);
   assert(x.ndim() == 2);
   int64_t n = x.shape(0);
   int64_t c = x.shape(1);
   assert(n == y.size());
-  float loss = 0;
+  double loss = 0;
   for (int64_t i = 0; i < n; i++) {
-    float max = x.at(i, 0);
+    double max = x.at(i, 0);
     for (int64_t j = 1; j < c; j++) {
-      float v = x.at(i, j);
+      double v = x.at(i, j);
       if (max < v) {
         max = v;
       }
     }
-    float sum = 0;
+    double sum = 0;
     for (int64_t j = 0; j < c; j++) {
       sum += dx->at(i, j) = std::exp(x.at(i, j) - max);
     }
